added Chapter 2 updates

+++ Chapter02/casting.cpp
@@ -3,12 +3,11 @@
 using namespace std;
 
 int main() {
-  
   int val1 = 100;
   int val2 = 23;
 
-  double result = ((int)val1 / (double)val2);
-  
+  double result = static_cast<double>(val1) / static_cast<double>(val2);
+
   cout << result << endl;
 
   return 0;
